+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- GET `/collections` collection search structured filtering extension with support for both CQL2 JSON and CQL2 text formats. [#471](https://github.com/stac-utils/stac-fastapi-elasticsearch-opensearch/pull/471)
+
 ### Changed
 
 ### Fixed